Add utilities to forward declare compound units
The mechanism may be a little bit complicated, but I think it's
manageable.  It's also necessary in order to get forward declaration
files working in Aurora's monorepo, which was what I used to design the
feature.  The fact that we're now able to do so gives me confidence that
Au's forward declarations now have everything they need to be useful and
usable in the real world.

I also added a how-to doc explaining how to use the forward declaration
feature generally.

Fixes #232.