Upgrade python packages (#345)
This fixes dependabot alerts 11, 12, and 13.  Granted, dependabot is
generally pretty meaningless for our use case, but (a) it's nice to get
back to a clean state, and (b) it's a good reminder to update our python
packages periodically.

I read the release notes for mike, mkdocs, and mkdocs-material.  The
former two had only bugfixes.  The latter had a new version of mermaid,
so I made sure to include pages that used mermaid diagrams in my A/B
blink comparisons.  It looks like it switched the order and changed the
sizes some, but the diagrams are equivalent and both are fine.

Before:


![image](https://github.com/user-attachments/assets/e38c4e39-c112-40ff-8e8c-a2d6cbe2445f)

After:


![image](https://github.com/user-attachments/assets/8eb99d0c-64ed-4696-8c86-e09d094aea0f)

Other than this, I really didn't see any difference at all in the A/B
comparison.

I also checked `mike` according to the instructions in
`REQUIREMENTS_LOCK.md`.  Still works!