Tighten NTTP creation: require `to_nttp(...)`
This backpedals on the very low friction pathway to pass `Quantity`
instances as template parameters.  This causes problems with C++20's
spaceship operator `<=>`.  We could fix this with a special C++20
compatibility layer, and we might still do exactly that.  However, this
is just one instance of a broader class of vulnerability.

It turns out that giving the very safe `Quantity` an **implicit**
conversion to a raw numeric type may simply be inviting trouble.  This
gives the compiler a pathway to generate all kinds of operations that we
probably don't want to generate.

We _could_ just make the conversion _explicit_... but then, how would we
trigger it?  I can't think of a way that would be at all comparably
concise to `to_nttp`.

So let's just use `to_nttp` for now.  It will be a lot easier to loosen
this later if we're confident that's correct, than to tighten it after
we've cut an official release.

Helps #353: it gives a more sensible error.