Provide compatibility with `<=>` in C++20 mode (#354)
This is the end of an era: the first Au core library code that uses a
macro.  Still, the use case meets the high bar we would need to justify
this: it keeps the library working out of the box for C++20 users,
without their needing to do anything.

To test this, we add a new C++20 mode to our CI tests.  We'll make sure
that it's working by pushing a first commit that should fail, because
all we have done is to add the test and run it in CI.  After that, we'll
push up the fix, and take the PR out of draft mode.

Fixes #353.  We may or may not also follow up to remove the implicit
conversion to the NTTP type, depending on the outcome of an internal
design discussion.