Improve error message for `Quantity` in unit slot (#366)
This is an easy mistake to make for users who haven't yet grasped the
idioms of the library.  Now, if they do this, they'll get a nice,
readable error, which directs them to the new section of the
troubleshooting doc (also added in this PR) that explains what is going
wrong, and what to do about it.

Fixes #327.