Add script to compare core bazel and CMake headers
This gives us the most important portion of #255.  In fact, it's even
better than what we originally suggested, which was a blocking checkbox
in code review.  This solution uses `bazel query` to get the
authoritative set of headers included in our main targets, `//au` and
`//au:io`.  It compares them to the CMake target `au`, which encompasses
both of these targets.

The script returns an exit code of `0` if the headers are an exact
match, and `1` otherwise.  If any files differ, it prints out the list
of every file that is found in only bazel or only CMake.

A good way to test this is to run:

```sh
check-cmake-headers && echo $?
```

This should print `0` on the last line.  Then, if we edit
`au/code/au/CMakeLists.txt`, and _change the name_ of one of the files
(say, prepend `asdf`), we should see the discrepancy printed out, and
the printed exit code should change to `1`.

A future PR will add a GitHub action to call this script and gate PRs on
its result.