Support streaming output for more types (#380)
We add support for `Magnitude`, `Constant`, and `SymbolFor`.

In the test, including `speed_of_light.hh` felt easier than defining a
new version of the speed of light constant.  This brought in the
authoritative definition of `Seconds`, so we deleted our ad hoc
definition.  In general, it would be nice to clearly articulate a policy
as to which header files we want to let ourselves use in which test
files: there's a tension between convenience, and keeping dependencies
clean.  But this is fine for now.

Fixes #378.