Resolve all `-Wsign-conversion` errors (#394)
The goal was to get this command to pass cleanly:

```sh
bazel build \
  --copt="-Werror" \
  --copt="-Wsign-conversion" \
  --config=gcc10 \
  //release/...:all \
  //au/...:all
```

The biggest change was to `string_constant`: we needed to fix up our
strategy for taking the absolute value of a number.  It was cleanest to
write a new helper, `abs_as_unsigned(T)`, which carefully and
efficiently handles all values (even the troublesome minimum
representable value, _whose negative can't be represented in `T`_). So
we're not just becoming warning-clean here; we're also handling values
that we couldn't handle before.  I found this cleaned up the
implementation a bit, too (no more `x = -x;`; no more ternary).

All of these new utilities got unit tests, to build confidence.

For the `unit_of_measure.hh` change, the right move was to change the
interface to accept `std::intmax` directly.  After all, literally the
only thing we do with `Deg` is to pass it to `UnitPowerT`, which takes
`std::intmax_t`.

For `magnitude.hh`, we simply insert an explicit static cast.  This is
very safe in this case, because `power` is the _log_ of a representable
number: it simply can't be very big at all.

Helps #392.  After this PR lands, #393 should become landable.

---------

Co-authored-by: Michael Hordijk <hordijk@aurora.tech>