Add "football fields" unit of measure (#401)
At the time of Aurora's successful Commercial Launch, delivering the
first NVO trucks regularly hauling loads on public roads, several press
releases and public news articles referred to the forward lidar sensor
range as "over four football fields".  It is clearly important to be
able to verify this measurement in C++ code.  Therefore, we add support
for this classic American unit of measure to Au.

To make sure we got it right, we compare not just against the
[authoritative] definition, but also against an Aurora [blog] which
gives "over 400 meters" as a reference point.

[authoritative]: https://en.wikipedia.org/wiki/American_football_field
[blog]: https://blog.aurora.tech/progress/firstlight-lidar-on-a-chip

---------

Co-authored-by: Michael Hordijk <hordijk@aurora.tech>