Handle negative magnitudes in conversions (#403)
Our code was all written under the assumption that a magnitude was
always positive.  Now we need to go back and see what changes when
magnitudes can be negative.

For applying magnitudes to values, the _category_ tends not to change.
In fact, for everything except the "apply rational magnitudes to
integers" case, the code is identical, and all we do is add more test
cases.  For that rational/integer case, though, we need to update our
checks for whether the numerator can fit: it's specifically the
_absolute value_ of the numerator that matters.  Additionally, we "short
circuit" the max non-overflowing value computation whenever the
magnitude is negative and the _destination_ is unsigned.  (Someday, once
we do #349, _all_ "apply magnitude" cases will have this form, but for
now the rational/integer case is special.)

On the conversion policy code, we now use the absolute value to check
for overflow.  We also add a check for setting an unsigned destination
with a negative scale factor, which can sometimes be OK, but is
certainly not OK to do _implicitly_.

Regarding common units, we used to consider a unit redundant if it was
an integer multiple of another.  Now that we can have negative integer
multiples, this is not what we want.  Consider "neginches" and "feet":
the core algorithm of "take the lowest power of every base" means that
for the base (-1), we should take the power 0 (positive input) rather
than the power 1 (negative input).  So the common unit should be
(something quantity-equivalent to) _inches_, and we shouldn't
short-circuit and return neginches just because every unit is a
(sometimes negative) integer multiple of it.

Finally, this comes with a new capability for `Quantity`: if the scale
factor is exactly `-1`, we support implicit conversions without checking
for overflow.  We add a test to cover this.

Helps #370.