Compare negative units correctly (#406)
For a `Quantity` of a negative unit, we compare by passing the arguments
in the reverse order.  This works for all equality and inequality
comparisons.

I kept the `using Twc` line next to the `operator<=>` line, instead of
with the other `using` statements, to reduce the number of instances of
`#if defined(__cpp_impl_three_way_comparison)...` we needed.

Helps #370.