Add `ComputeOriginDisplacementUnit` helper
This will produce an actual unit whenever possible --- that is, whenever
the input units have the same dimension, but different origins.
Otherwise, it will produce `Zero`.  Of course, in the latter case, this
isn't an actual "unit", but it will participate correctly in the common
unit machinery.

We automatically generate a label, which is consistent with our
(somewhat quirky and dubious) `@(...)` syntax for printing
`QuantityPoint`.

Helps #369.