Simplify common point unit implementation (#421)
Now, the common point unit is just the _common unit_, except that
instead of just looking at all of the input units, we also consider
their mutual origin displacements (treated as new ad hoc units).  And,
we set an origin that is equal to the lowest origin among those inputs.
That's it!  No more ugly hacks... goodbye `TypeHoldingCommonOrigin`!

This simplifies our logic for conversion checking in `QuantityPoint`,
and maximally expands the set of allowed operations.  This is made
possible because the origin displacement is now always a shapeshifter
type, which has a perfect conversion policy, known at compile time, to
every destination type.  I believe this is a first among units
libraries.

Some people might have been using the `OriginDisplacement` trait, so we
retain it, but it has to move from `unit_of_measure.hh` to
`quantity_point.hh`, because now it depends on `:constant`.  We document
this.

We did need to refine the ordering by adding another tiebreaker based on
the origin displacement unit, too.

Fixes https://github.com/aurora-opensource/au/issues/369.

---------

Co-authored-by: Michael Hordijk <hordijk@aurora.tech>