Rename `detail` namespace to `auimpl`
I think a "bite the bullet megadiff" makes the most sense here, because
the changes should all be the same length, so the PR should be fairly
easy to review.  The reviewer can check the Viewed box for one file at a
time, and thus not be forced to do it all in one pass --- as long as we
do end up landing this relatively soon before we try other changes.