Implement `lerp` (#430)
This wraps `std::lerp`.  Includes documentation.

The main implementations, in `math.hh`, convert each input to the common
unit before performing the calculation, and then wrap the result.

The hidden friend implementation, in `Quantity`, is necessary to support
shapeshifter types such as `Zero` and `Constant<U>`.  We would need to
add one in `QuantityPoint` too in order to support shapeshifter
overloads, but `QuantityPoint` doesn't have any shapeshifter types in
the current design.

Fixes #390.