Add traits for sign of units and magnitudes
The "sign" of a unit or magnitude is always `mag<1>()` (i.e.,
`Magnitude<>{}`) if positive, or `-mag<1>()` (i.e.,
`Magnitude<Negative>{}`) if negative.

We also update the docs, including adding docs for a few new traits that
we had missed, and updating the headings for the magnitude traits to
phrase the boolean ones in the form of a question (in case we ever get
on Jeopardy).

Helps #428.