Compare underlying values of hetereogeneous types (#432)
Formerly, in `Quantity`, we used a `CompareUnderlyingValues` helper,
and a `UnitSign` enum.  The helper was responsible for taking two
_identical_ `Quantity` types, and extracting their values.  (We built
out this machinery to support negative units.)

Now, we refactor to use a new helper, `SignAwareComparison`.  This
brings two simplifications and one generalization.

- **Simplification:** stop doing the extraction in the helper; let the
  callers extract and pass the values.  (This gives the helper just one
  responsibility: namely, take the unit sign into account when deciding
  which order to pass the parameters.)

- **Simplification:** lose the enum; just use the new `UnitSign` trait.

- **Generalization:** accept _heterogeneous_ parameters.

This last change will allow us to detect when we are passing mixed
signed and unsigned integral arguments.  Of course, this PR is a no-op
refactor, so we're not actually taking advantage of this yet.  That will
come in a follow-on PR.

Helps #428.

---------

Co-authored-by: Michael Hordijk <hordijk@aurora.tech>